mario.c
A function that takes a numerical input and creates a pyramid in the same style as old Mario level end screens before Mario jumps onto the flag at the end of the level. The pyramid will be (input) high and (input) wide at the base, oriented to the right so that the player can jump up the pyramid blocks like stairs.